// src/types/supabase.d.ts
export type Json = string | number | boolean | null | { [key: string]: Json | undefined } | Json[];

export interface Database {
  public: {
    Tables: {
// ...
      bookmarks: {
        Row: {
          id: string;
          user_id: string;
          post_id: string;
          title?: string;
          slug?: string;
          excerpt?: string | null;
          featured_image?: Json | null;
          category?: string | null;
          tags?: string[] | null;
          read_status?: 'read' | 'unread' | null;
          notes?: string | null;
          created_at: string;
          collection_id?: string | null;
        };
        Insert: {
          id?: string;
          user_id: string;
          post_id: string;
          title?: string;
          slug?: string;
          excerpt?: string | null;
          featured_image?: Json | null;
          category?: string | null;
          tags?: string[] | null;
          read_status?: 'read' | 'unread' | null;
          notes?: string | null;
          created_at?: string;
          collection_id?: string | null;
        };
        Update: {
          id?: string;
          user_id?: string;
          post_id?: string;
          title?: string;
          slug?: string;
          excerpt?: string | null;
          featured_image?: Json | null;
          category?: string | null;
          tags?: string[] | null;
          read_status?: 'read' | 'unread' | null;
          notes?: string | null;
          created_at?: string;
          collection_id?: string | null;
        };
      };
// ...
    };
    Views: {
      [_ in never]: never;
    };
    Functions: {
      [_ in never]: never;
    };
    Enums: {
      [_ in never]: never;
    };
  };
}